url/host-parser.html

<!doctype html>
<meta charset=utf-8>
<title>Testing the host parser</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div id=log></div>
<script>
let setup = async_test("Loading data…")
setup.step(() => {
  let request = new XMLHttpRequest()
  request.open("GET", "hosttestdata.json")
  request.send()
  request.responseType = "json"
  request.onload = setup.step_func_done(() => runHostTests(request.response) )
})

function runHostTests(hostTests) {
  for(var i = 0, l = hostTests.length; i < l; i++) {
    let hostTest = hostTests[i]
    if (typeof hostTest === "string") {
      continue // skip comments
    }

    if(!hostTest.onlySetters) {
      test(() => {
        if(hostTest.output !== null) {
          const url = new URL("https://" + hostTest.input + "/x")
          assert_equals(url.pathname, "/x")
          assert_equals(url.host, hostTest.output)
          assert_equals(url.hostname, hostTest.output)
        } else {
          assert_throws(new TypeError, () => new URL("https://" + hostTest.input + "/x"))
        }
      }, encodeURI(hostTest.input))
    }

    test(() => {
      const url = new URL("https://x")
      ;["host", "hostname"].forEach((val) => {
        url[val] = hostTest.input
        if(hostTest.output !== null) {
          assert_equals(url[val], hostTest.output)
        } else {
          assert_equals(url[val], "x")
        }
      })
    }, "host/hostname setter for " + encodeURI(hostTest.input))

    // TODO: also test <a> and <area>
  }
}
</script>
